+++ tests/src/rules/dynamic-import-chunkname.js
@@ -1,5 +1,6 @@
-import { SYNTAX_CASES } from '../utils'
+import { SYNTAX_CASES, getTSParsers } from '../utils'
 import { RuleTester } from 'eslint'
+import semver from 'semver'
 
 const rule = require('rules/dynamic-import-chunkname')
 const ruleTester = new RuleTester()
@@ -20,8 +21,8 @@ const noLeadingCommentError = 'dynamic imports require a leading comment with th
 const nonBlockCommentError = 'dynamic imports require a /* foo */ style comment, not a // foo comment'
 const noPaddingCommentError = 'dynamic imports require a block comment padded with spaces - /* foo */'
 const invalidSyntaxCommentError = 'dynamic imports require a "webpack" comment with valid syntax'
-const commentFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: "${commentFormat}",? */`
-const pickyCommentFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: "${pickyCommentFormat}",? */`
+const commentFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: ["']${commentFormat}["'],? */`
+const pickyCommentFormatError = `dynamic imports require a leading comment in the form /* webpackChunkName: ["']${pickyCommentFormat}["'],? */`
 
 ruleTester.run('dynamic-import-chunkname', rule, {
   valid: [
@@ -133,6 +134,14 @@ ruleTester.run('dynamic-import-chunkname', rule, {
     },
     {
       code: `import(
+        /* webpackChunkName: 'someModule' */
+        'someModule'
+      )`,
+      options,
+      parser,
+    },
+    {
+      code: `import(
         /* webpackChunkName: "someModule" */
         'someModule'
       )`,
@@ -191,17 +200,33 @@ ruleTester.run('dynamic-import-chunkname', rule, {
     },
     {
       code: `import(
-        /* webpackChunkName: 'someModule' */
+        /* webpackChunkName: "someModule' */
         'someModule'
       )`,
       options,
       parser,
       output: `import(
-        /* webpackChunkName: 'someModule' */
+        /* webpackChunkName: "someModule' */
         'someModule'
       )`,
       errors: [{
-        message: commentFormatError,
+        message: invalidSyntaxCommentError,
+        type: 'CallExpression',
+      }],
+    },
+    {
+      code: `import(
+        /* webpackChunkName: 'someModule" */
+        'someModule'
+      )`,
+      options,
+      parser,
+      output: `import(
+        /* webpackChunkName: 'someModule" */
+        'someModule'
+      )`,
+      errors: [{
+        message: invalidSyntaxCommentError,
         type: 'CallExpression',
       }],
     },
@@ -422,21 +447,6 @@ ruleTester.run('dynamic-import-chunkname', rule, {
     },
     {
       code: `dynamicImport(
-        /* webpackChunkName: 'someModule' */
-        'someModule'
-      )`,
-      options,
-      output: `dynamicImport(
-        /* webpackChunkName: 'someModule' */
-        'someModule'
-      )`,
-      errors: [{
-        message: commentFormatError,
-        type: 'CallExpression',
-      }],
-    },
-    {
-      code: `dynamicImport(
         /* webpackChunkName "someModule" */
         'someModule'
       )`,
@@ -481,4 +491,333 @@ ruleTester.run('dynamic-import-chunkname', rule, {
       }],
     },
   ],
+})
+
+context('TypeScript', () => {
+  getTSParsers().forEach((typescriptParser) => {
+    const nodeType = typescriptParser.includes('typescript-eslint-parser') || (typescriptParser.includes('@typescript-eslint/parser') && semver.satisfies(require('@typescript-eslint/parser/package.json').version, '^2'))
+      ? 'CallExpression'
+      : 'ImportExpression'
+
+    ruleTester.run('dynamic-import-chunkname', rule, {
+      valid: [
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "Some_Other_Module" */
+            "test"
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "SomeModule123" */
+            "test"
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule", webpackPrefetch: true */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule", webpackPrefetch: true, */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackPrefetch: true, webpackChunkName: "someModule" */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackPrefetch: true, webpackChunkName: "someModule", */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackPrefetch: true */
+            /* webpackChunkName: "someModule" */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            /* webpackPrefetch: true */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options: pickyCommentOptions,
+          parser: typescriptParser,
+          errors: [{
+            message: pickyCommentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName: 'someModule' */
+            'test'
+          )`,
+          options,
+          parser: typescriptParser,
+        },
+      ],
+      invalid: [
+        {
+          code: `import(
+            // webpackChunkName: "someModule"
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            // webpackChunkName: "someModule"
+            'someModule'
+          )`,
+          errors: [{
+            message: nonBlockCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: 'import(\'test\')',
+          options,
+          parser: typescriptParser,
+          output: 'import(\'test\')',
+          errors: [{
+            message: noLeadingCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName: someModule */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName: someModule */
+            'someModule'
+          )`,
+          errors: [{
+            message: invalidSyntaxCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName "someModule' */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName "someModule' */
+            'someModule'
+          )`,
+          errors: [{
+            message: invalidSyntaxCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName 'someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName 'someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: invalidSyntaxCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: invalidSyntaxCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName:"someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName:"someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /*webpackChunkName: "someModule"*/
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /*webpackChunkName: "someModule"*/
+            'someModule'
+          )`,
+          errors: [{
+            message: noPaddingCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName  :  "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName  :  "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule" ; */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName: "someModule" ; */
+            'someModule'
+          )`,
+          errors: [{
+            message: invalidSyntaxCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* totally not webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* totally not webpackChunkName: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: invalidSyntaxCommentError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackPrefetch: true */
+            /* webpackChunk: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackPrefetch: true */
+            /* webpackChunk: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackPrefetch: true, webpackChunk: "someModule" */
+            'someModule'
+          )`,
+          options,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackPrefetch: true, webpackChunk: "someModule" */
+            'someModule'
+          )`,
+          errors: [{
+            message: commentFormatError,
+            type: nodeType,
+          }],
+        },
+        {
+          code: `import(
+            /* webpackChunkName: "someModule123" */
+            'someModule'
+          )`,
+          options: pickyCommentOptions,
+          parser: typescriptParser,
+          output: `import(
+            /* webpackChunkName: "someModule123" */
+            'someModule'
+          )`,
+          errors: [{
+            message: pickyCommentFormatError,
+            type: nodeType,
+          }],
+        },
+      ],
+    })
+  })
 })