+++ tests/src/rules/export.js
@@ -1,12 +1,14 @@
-import { test, testFilePath, SYNTAX_CASES, getTSParsers } from '../utils'
+import { test, testFilePath, SYNTAX_CASES, getTSParsers, testVersion } from '../utils'
 
 import { RuleTester } from 'eslint'
+import eslintPkg from 'eslint/package.json'
+import semver from 'semver'
 
 var ruleTester = new RuleTester()
   , rule = require('rules/export')
 
 ruleTester.run('export', rule, {
-  valid: [
+  valid: [].concat(
     test({ code: 'import "./malformed.js"' }),
 
     // default
@@ -24,7 +26,25 @@ ruleTester.run('export', rule, {
     test({ code: 'export default foo; export * from "./bar"' }),
 
     ...SYNTAX_CASES,
-  ],
+
+    test({
+      code: `
+        import * as A from './named-export-collision/a';
+        import * as B from './named-export-collision/b';
+
+        export { A, B };
+      `,
+    }),
+    testVersion('>= 6', () => ({
+      code: `
+        export * as A from './named-export-collision/a';
+        export * as B from './named-export-collision/b';
+      `,
+      parserOptions: {
+        ecmaVersion: 2020,
+      },
+    })) || [],
+  ),
 
   invalid: [
     // multiple defaults
@@ -191,6 +211,16 @@ context('TypeScript', function () {
           code: 'export * from "./file1.ts"',
           filename: testFilePath('typescript-d-ts/file-2.ts'),
         }, parserConfig)),
+
+        ...(semver.satisfies(eslintPkg.version, '< 6') ? [] : [
+          test({
+            code: `
+              export * as A from './named-export-collision/a';
+              export * as B from './named-export-collision/b';
+            `,
+            parser: parser,
+          }),
+        ]),
       ],
       invalid: [
         // type/value name clash