+++ tests/src/rules/no-extraneous-dependencies.js
@@ -1,11 +1,15 @@
-import { test } from '../utils'
-import * as path from 'path'
-import * as fs from 'fs'
+import { getTSParsers, test, testFilePath } from '../utils'
+import typescriptConfig from '../../../config/typescript'
+import path from 'path'
+import fs from 'fs'
+import semver from 'semver'
+import eslintPkg from 'eslint/package.json'
 
 import { RuleTester } from 'eslint'
 import flatMap from 'array.prototype.flatmap'
 
 const ruleTester = new RuleTester()
+const typescriptRuleTester = new RuleTester(typescriptConfig)
 const rule = require('rules/no-extraneous-dependencies')
 
 const packageDirWithSyntaxError = path.join(__dirname, '../../files/with-syntax-error')
@@ -17,6 +21,7 @@ const packageFileWithSyntaxErrorMessage = (() => {
   }
 })()
 const packageDirWithFlowTyped = path.join(__dirname, '../../files/with-flow-typed')
+const packageDirWithTypescriptDevDependencies = path.join(__dirname, '../../files/with-typescript-dev-dependencies')
 const packageDirMonoRepoRoot = path.join(__dirname, '../../files/monorepo')
 const packageDirMonoRepoWithNested = path.join(__dirname, '../../files/monorepo/packages/nested-package')
 const packageDirWithEmpty = path.join(__dirname, '../../files/empty')
@@ -312,3 +317,74 @@ ruleTester.run('no-extraneous-dependencies', rule, {
     }),
   ],
 })
+
+describe('TypeScript', function () {
+  getTSParsers().forEach((parser) => {
+    const parserConfig = {
+      parser: parser,
+      settings: {
+        'import/parsers': { [parser]: ['.ts'] },
+        'import/resolver': { 'eslint-import-resolver-typescript': true },
+      },
+    }
+
+    if (parser !== require.resolve('typescript-eslint-parser')) {
+      ruleTester.run('no-extraneous-dependencies', rule, {
+        valid: [
+          test(Object.assign({
+            code: 'import type { JSONSchema7Type } from "@types/json-schema";',
+            options: [{packageDir: packageDirWithTypescriptDevDependencies, devDependencies: false }],
+          }, parserConfig)),
+        ],
+        invalid: [
+          test(Object.assign({
+            code: 'import { JSONSchema7Type } from "@types/json-schema";',
+            options: [{packageDir: packageDirWithTypescriptDevDependencies, devDependencies: false }],
+            errors: [{
+              message: "'@types/json-schema' should be listed in the project's dependencies, not devDependencies.",
+            }],
+          }, parserConfig)),
+        ],
+      })
+    } else {
+      ruleTester.run('no-extraneous-dependencies', rule, {
+        valid: [],
+        invalid: [
+          test(Object.assign({
+            code: 'import { JSONSchema7Type } from "@types/json-schema";',
+            options: [{packageDir: packageDirWithTypescriptDevDependencies, devDependencies: false }],
+            errors: [{
+              message: "'@types/json-schema' should be listed in the project's dependencies, not devDependencies.",
+            }],
+          }, parserConfig)),
+          test(Object.assign({
+            code: 'import type { JSONSchema7Type } from "@types/json-schema";',
+            options: [{packageDir: packageDirWithTypescriptDevDependencies, devDependencies: false }],
+            errors: [{
+              message: "'@types/json-schema' should be listed in the project's dependencies, not devDependencies.",
+            }],
+          }, parserConfig)),
+        ],
+      })
+    }
+  })
+})
+
+if (semver.satisfies(eslintPkg.version, '>5.0.0')) {
+  typescriptRuleTester.run('no-extraneous-dependencies typescript type imports', rule, {
+    valid: [
+      test({
+        code: 'import type MyType from "not-a-dependency";',
+        filename: testFilePath('./no-unused-modules/typescript/file-ts-a.ts'),
+        parser: require.resolve('babel-eslint'),
+      }),
+      test({
+        code: 'import type { MyType } from "not-a-dependency";',
+        filename: testFilePath('./no-unused-modules/typescript/file-ts-a.ts'),
+        parser: require.resolve('babel-eslint'),
+      }),
+    ],
+    invalid: [
+    ],
+  })
+}