+++ tests/src/rules/order.js
@@ -711,6 +711,78 @@ ruleTester.run('order', rule, {
         },
       ],
     }),
+    ...flatMap(getTSParsers, parser => [
+      // Order of the `import ... = require(...)` syntax
+      test({
+        code: `
+          import blah = require('./blah');
+          import { hello } from './hello';`,
+        parser,
+        options: [
+          {
+            alphabetize: {
+              order: 'asc',
+            },
+          },
+        ],
+      }),
+      // Order of object-imports
+      test({
+        code: `
+          import blah = require('./blah');
+          import log = console.log;`,
+        parser,
+        options: [
+          {
+            alphabetize: {
+              order: 'asc',
+            },
+          },
+        ],
+      }),
+      // Object-imports should not be forced to be alphabetized
+      test({
+        code: `
+          import debug = console.debug;
+          import log = console.log;`,
+        parser,
+        options: [
+          {
+            alphabetize: {
+              order: 'asc',
+            },
+          },
+        ],
+      }),
+      test({
+        code: `
+          import log = console.log;
+          import debug = console.debug;`,
+        parser,
+        options: [
+          {
+            alphabetize: {
+              order: 'asc',
+            },
+          },
+        ],
+      }),
+      test({
+        code: `
+          import { a } from "./a";
+          export namespace SomeNamespace {
+              export import a2 = a;
+          }
+        `,
+        parser,
+        options: [
+          {
+            groups: ['external', 'index'],
+            alphabetize: { order: 'asc' },
+          },
+        ],
+      }),
+    ]),
   ],
   invalid: [
     // builtin before external module (require)
@@ -1167,6 +1239,7 @@ ruleTester.run('order', rule, {
       }],
     }),
     ...flatMap(getTSParsers(), parser => [
+      // Order of the `import ... = require(...)` syntax
       test({
         code: `
           var fs = require('fs');
@@ -1183,7 +1256,7 @@ ruleTester.run('order', rule, {
           message: '`fs` import should occur after import of `../foo/bar`',
         }],
       }),
-      {
+      test({
         code: `
           var async = require('async');
           var fs = require('fs');
@@ -1196,7 +1269,7 @@ ruleTester.run('order', rule, {
         errors: [{
           message: '`fs` import should occur before import of `async`',
         }],
-      },
+      }),
       test({
         code: `
           import sync = require('sync');
@@ -1217,6 +1290,16 @@ ruleTester.run('order', rule, {
         parser,
         errors: [{
           message: '`async` import should occur before import of `sync`',
+        }],
+      }),
+      // Order of object-imports
+      test({
+        code: `
+          import log = console.log;
+          import blah = require('./blah');`,
+        parser,
+        errors: [{
+          message: '`./blah` import should occur before import of `console.log`',
         }],
       }),
     ]),